Align secrets list page with prototype header design
